requirement_windows.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
idna==3.6
imageio==2.33.1
imbalanced-learn==0.12.0
imblearn==0.0
Jinja2==3.1.3
joblib==1.2.0
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.4
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
pillow==10.2.0
psycopg2==2.9.9
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.4
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.12.0
seaborn==0.13.2
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
sympy==1.12
threadpoolctl==3.2.0
tifffile==2024.1.30
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
